move ClickTarget code to `mod vector::click_target`
