+++ editor/src/messages/tool/common_functionality/shapes/shape_utility.rs
@@ -13,7 +13,7 @@ use bezier_rs::Subpath;
 use glam::{DAffine2, DMat2, DVec2};
 use graph_craft::document::NodeInput;
 use graph_craft::document::value::TaggedValue;
-use graphene_std::renderer::ClickTargetType;
+use graphene_std::vector::click_target::ClickTargetType;
 use graphene_std::vector::misc::dvec2_to_point;
 use kurbo::{BezPath, PathEl, Shape};
 use std::collections::VecDeque;
