+++ node-graph/gcore/src/vector/mod.rs
@@ -1,5 +1,6 @@
 mod algorithms;
 pub mod brush_stroke;
+pub mod click_target;
 pub mod generator_nodes;
 pub mod misc;
 mod reference_point;
